+++ src/components/header/sign-out-link.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import { Link } from 'react-router-dom';
 import { AppRoutes } from '../../const';
 import { useAppDispatch } from '../../hooks/use-store';
-import { logoutAction } from '../../store/api-action';
+import { logoutAction } from '../../store/api-actions/user';
 
 const SignOutLink : FC = () => {
   const dispatch = useAppDispatch();
