+++ src/components/other/start-fetch.tsx
@@ -1,6 +1,8 @@
 import { FC, PropsWithChildren } from 'react';
 import { useAppDispatch } from '../../hooks/use-store';
-import { checkAuthAction, fetchOffersAction, fetchFavoritesAction } from '../../store/api-action';
+import { checkAuthAction } from '../../store/api-actions/user';
+import { fetchFavoritesAction } from '../../store/api-actions/favorite';
+import { fetchOffersAction } from '../../store/api-actions/offer';
 
 const StartFetch : FC<PropsWithChildren> = ({children}) => {
   const dispatch = useAppDispatch();
